requirements.txt
flask
flask-cors
pymongo
python-dotenv
ultralytics
opencv-python-headless
gunicorn